% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseInt.R
\name{doseInt}
\alias{doseInt}
\title{Fits dose interval}
\usage{
doseInt(y = NULL, x = NULL, a = NULL, s = NULL, propensity = NULL,
  train = NULL, alpha = c(0.5, 0.5), type = "PDI",
  family = c("continuous", "ordinal", "as.ordinal"), two.sided = FALSE,
  method = c("rq", "svmLinear", "svmRadial", "RLT", "tree"), pred0 = NULL,
  maxiter = 20, step = 1, trace = 0, lower = TRUE, lambda = NULL,
  cost = 1, embed.mtry = 1/2, K = 10, global = F, margin = 0,
  breaks = "quantile", ...)
}
\arguments{
\item{train}{the training data set.}

\item{alpha}{the guaranteed probability in PDI and the relative loss in EDI. Default is c(0.5,0.5).}

\item{type}{indicates whether PDI or EDI is to be calculated.}

\item{family}{specifis the methods. 'continuous' leads to DC-algorithm,'ordinal' handles the ordinal treatments, while 'as.ordinal' cuts continuous treatment into several ordinal levels and apply the ordinal algorithm.}

\item{two.sided}{indicator of whether two-sided interval is considered.}

\item{method}{specified methods used for DC and ordinal approaches.}

\item{pred0}{the initial prediction for training data set, if NULL, default initialization is used}

\item{maxiter}{maximal iterations for DC algorithm}

\item{step}{stepsize for DC algorithm}

\item{trace}{the level of detail to be printed: 1,2,3,4}

\item{lower}{True or False, Whether lower boundary is considered when applied to one-sided interval}

\item{K}{level of ordinal treatment, default 10}

\item{global}{if global data are used for ordinal approach, default False}

\item{margin}{how large is the margin if global=False, default 0}

\item{breaks}{ways 'as.ordinal' cuts continuous treatment into several ordinal levels, default 'quantile'. User can also choose 'uniform' or specify customized breaks}

\item{test}{the testing data set. If test is NULL, use train as test.}

\item{nfolds}{the number of folds used in cross-validation.}

\item{Eps}{parameter for the relaxation}

\item{Lambda}{penalty for the quantile regression}

\item{Cost}{cost for the support vector machine}

\item{Embed.mtry}{proportion of mtry}
}
\description{
This function is the wrapper of fitting functions for doseInt
}
\examples{
train1=Scenario2.continuous(2000,50,1)
test1=Scenario2.continuous(10000,50,1)
fit=doseInt(train=train1,
            alpha=c(0.5,0.5),type='PDI',
            family='continuous',
            two.sided=FALSE,
            method=c('rq','svmRadial'),
            maxiter=20,step=1,trace=3,lower=TRUE,lambda=0.000001,
            K=10,global=F,margin=0)
pred=predict(fit,test1)
}
