% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{find.ordinal}
\alias{find.ordinal}
\title{Fits dose interval using ordinal approach}
\usage{
find.ordinal(x)
}
\arguments{
\item{train}{the training data set.}

\item{test}{the testing data set. If test is NULL, use train as test.}

\item{pred0}{the initial prediction for training data set, if NULL, default initialization is used}

\item{nfolds}{the number of folds used in cross-validation.}

\item{alpha}{the guaranteed probability in PDI and the relative loss in EDI. Default is c(0.5,0.5).}

\item{type}{indicates whether PDI or EDI is to be calculated.}

\item{two.sided}{indicator of whether two-sided interval is considered.}

\item{family}{specifis the methods. 'continuous' leads to DC-algorithm,'ordinal' handles the ordinal treatments, while 'as.ordinal' cuts continuous treatment into several ordinal levels and apply the ordinal algorithm.}

\item{method}{specified methods used for DC and ordinal approaches.}

\item{trace}{the level of detail to be printed: 1,2,3,4}

\item{lower}{True or False, Whether lower boundary is considered when applied to one-sided interval}

\item{K}{level of ordinal treatment, default 10}

\item{global}{if global data are used for ordinal approach, default False}

\item{margin}{how large is the margin if global=False, default 0}

\item{breaks}{ways 'as.ordinal' cuts continuous treatment into several ordinal levels, default 'quantile'. User can also choose 'uniform' or specify customized breaks}

\item{Cost}{cost for the support vector machine}
}
\description{
This function is the wrapper of fitting functions for doseInt
}
\examples{



}
