% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{Scenario1.continuous}
\alias{Scenario1.continuous}
\title{Four scenarios for simulating dose interval data}
\usage{
Scenario1.continuous(size, ncov, seed)
}
\arguments{
\item{size}{the number of observations to be drawn}

\item{ncov}{the number of covariates to be generated}

\item{seed}{random seed}
}
\description{
there are 4 scenarios: linear one-sided, non-linear one-sided, linear two-sided and non-linear two-sided. Each scenario has 1 continuous treatment case and an ordinal treatment case.
}
\examples{

}
