% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.doseInt.R
\name{cv.doseInt}
\alias{cv.doseInt}
\title{Conduct cross-validation for dose interval}
\usage{
cv.doseInt(train, test = NULL, pred0 = NULL, nfolds = 5, alpha = c(0.5,
  0.5), type = "PDI", two.sided = FALSE, family = c("continuous",
  "ordinal", "as.ordinal"), method = c("rq", "svmLinear", "svmRadial", "RLT",
  "tree"), maxiter = 20, step = 1, trace = 0, lower = TRUE, K = 10,
  global = F, margin = 0, breaks = "quantile", Eps = c(0.1, 0.2, 0.3),
  Lambda = c(2^(-(0:10)), 1e-08), Cost = c(1e-05, 0.001, 0.1, seq(1, 15,
  2)), Embed.mtry = c(0.2, 0.3, 0.4, 0.5), ...)
}
\arguments{
\item{train}{the training data set.}

\item{test}{the testing data set. If test is NULL, use train as test.}

\item{pred0}{the initial prediction for training data set, if NULL, default initialization is used}

\item{nfolds}{the number of folds used in cross-validation.}

\item{alpha}{the guaranteed probability in PDI and the relative loss in EDI. Default is c(0.5,0.5).}

\item{type}{indicates whether PDI or EDI is to be calculated.}

\item{two.sided}{indicator of whether two-sided interval is considered.}

\item{family}{specifis the methods. 'continuous' leads to DC-algorithm,'ordinal' handles the ordinal treatments, while 'as.ordinal' cuts continuous treatment into several ordinal levels and apply the ordinal algorithm.}

\item{method}{specified methods used for DC and ordinal approaches.}

\item{maxiter}{maximal iterations for DC algorithm}

\item{step}{stepsize for DC algorithm}

\item{trace}{the level of detail to be printed: 1,2,3,4}

\item{lower}{True or False, Whether lower boundary is considered when applied to one-sided interval}

\item{K}{level of ordinal treatment, default 10}

\item{global}{if global data are used for ordinal approach, default False}

\item{margin}{how large is the margin if global=False, default 0}

\item{breaks}{ways 'as.ordinal' cuts continuous treatment into several ordinal levels, default 'quantile'. User can also choose 'uniform' or specify customized breaks}

\item{Eps}{parameters for the relaxation}

\item{Lambda}{candidate penalties for the quantile regression}

\item{Cost}{candidate costs for the support vector machine}

\item{Embed.mtry}{candidate proportions of mtry}
}
\description{
This function is the wrapper of cross-validation for doseInt
}
\examples{



}
